% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.playground-package}
\alias{blockr.playground}
\alias{blockr.playground-package}
\title{blockr.playground: Highcharts Visualization Blocks for Blockr}
\description{
Extends 'blockr.core' with interactive blocks for data visualization using 'highcharter'. Users can build charts through a graphical interface without writing code directly. Includes common chart types (scatter, line, column, area, pie) with aesthetic mappings for color and size.
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item David Granjon \email{david@cynkra.com}
  \item Mike Page \email{mike@cynkra.com}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
