% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echart-candlestick-block.R
\name{new_echart_candlestick_block}
\alias{new_echart_candlestick_block}
\alias{block_ui.echart_candlestick_block}
\alias{block_output.echart_candlestick_block}
\alias{board_options.echart_candlestick_block}
\alias{block_render_trigger.echart_candlestick_block}
\title{ECharts Candlestick Block}
\usage{
new_echart_candlestick_block(
  date = character(),
  open = character(),
  close = character(),
  low = character(),
  high = character(),
  title = character(),
  theme = "default",
  ...
)

\method{block_ui}{echart_candlestick_block}(id, x, ...)

\method{block_output}{echart_candlestick_block}(x, result, session)

\method{board_options}{echart_candlestick_block}(x, ...)

\method{block_render_trigger}{echart_candlestick_block}(x, session = blockr.core::get_session())
}
\arguments{
\item{date}{Column for date/time x-axis}

\item{open}{Column for opening price}

\item{close}{Column for closing price}

\item{low}{Column for low price}

\item{high}{Column for high price}

\item{title}{Chart title}

\item{theme}{ECharts theme name}

\item{...}{Forwarded to \code{\link[blockr.core]{new_transform_block}}}

\item{id}{Module ID}

\item{x}{Block object}

\item{result}{Evaluation result}

\item{session}{Shiny session object}
}
\value{
A block object of class \code{echart_candlestick_block}.
}
\description{
A specialized block for creating candlestick chart visualizations with ECharts.
Used for displaying financial OHLC (Open, High, Low, Close) data.
}
\examples{
# Create a candlestick block
new_echart_candlestick_block(
  date = "date", open = "open", close = "close",
  low = "low", high = "high"
)

if (interactive()) {
  library(blockr.core)
  stock_data <- data.frame(
    date = as.Date("2024-01-01") + 0:9,
    open = c(100, 102, 101, 103, 102, 104, 103, 105, 104, 106),
    close = c(102, 101, 103, 102, 104, 103, 105, 104, 106, 105),
    low = c(99, 100, 100, 101, 101, 102, 102, 103, 103, 104),
    high = c(103, 103, 104, 104, 105, 105, 106, 106, 107, 107)
  )
  serve(new_echart_candlestick_block(), list(data = stock_data))
}

}
