% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{register_playground_blocks}
\alias{register_playground_blocks}
\title{Register playground blocks}
\usage{
register_playground_blocks()
}
\value{
NULL (invisibly)
}
\description{
Registers all blocks from the blockr.playground package.
}
